refactor: pages refactored
