+++ client/src/components/Room/CreateRoom/style.module.css
@@ -1,17 +1,3 @@
-.container {
-    /* margin: 5% 30%; */
-    max-width: 500px;
-    margin: 50px auto;
-    justify-content: center;
-}
-
-.outerCard {
-    background: inherit;
-    border-radius: 6px;
-    border: none;
-    box-shadow: 0 -2px 10px rgb(75, 120, 202);
-}
-
 .innerCard {
     background: inherit;
 }
