--- client/src/pages/JoinRoomPage.tsx
@@ -1,12 +1,0 @@
-import Room from 'components/Room';
-import React from 'react';
-
-const JoinRoomPage = () => {
-    return (
-        <>
-            <Room />
-        </>
-    );
-};
-
-export default JoinRoomPage;
