+++ client/src/pages/SnippetPage/SnippetPage.tsx
@@ -1,5 +1,6 @@
 import Snippet from 'components/Snippet';
 import React from 'react';
+
 const SnippetPage = () => {
     return (
         <>
