+++ client/src/pages/SnippetsPage/SnippetsPage.tsx
