+++ client/src/routing/routes.tsx
@@ -1,10 +1,10 @@
-import HomePage from 'pages/HomePage';
-import SnippetPage from 'pages/SnippetPage';
-import SnippetsPage from 'pages/SnippetsPage';
+import HomePage from 'pages/HomePage/HomePage';
+import SnippetPage from 'pages/SnippetPage/SnippetPage';
+import SnippetsPage from 'pages/SnippetsPage/SnippetsPage';
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import JoinRoomPage from 'pages/JoinRoomPage';
-import RoomPage from 'pages/RoomPage';
+import JoinRoomPage from 'pages/CollabPage/CollabPage';
+import RoomPage from 'pages/RoomPage/RoomPage';
 import RoomRoute from './RoomRoute';
 
 export const useRoutes = () => {
@@ -15,7 +15,7 @@ export const useRoutes = () => {
                 <Route exact path='/snippets' component={SnippetsPage} />
                 <Route exact path='/snippets/:id' component={SnippetPage} />
                 <Route exact path='/room' component={JoinRoomPage} />
-                <Route exact path='/room/:roomId' component={RoomPage} />
+                <RoomRoute exact path='/room/:roomId' component={RoomPage} />
             </Switch>
         </>
     );
